day1/requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
email-validator==2.1.0
WTForms==3.1.2
python-dotenv==1.0.1
Pillow==10.3.0
google-generativeai==0.3.2
itsdangerous==2.1.2
